+++ src/Physics.cpp
@@ -15,7 +15,7 @@ bool Physics::Vector2::operator!=(const Physics::Vector2 &v) const {
   return x != v.x || y != v.y;
 }
 
-Physics::Vector2 Physics::Vector2::operator+(const float f) const {
+Physics::Vector2 Physics::Vector2::operator+(const int f) const {
   return Physics::Vector2(x + f, y + f);
 }
 
@@ -23,7 +23,7 @@ Physics::Vector2 Physics::Vector2::operator+(const Physics::Vector2 &v) const {
   return Physics::Vector2(x + v.x, y + v.y);
 }
 
-Physics::Vector2 Physics::Vector2::operator-(const float f) const {
+Physics::Vector2 Physics::Vector2::operator-(const int f) const {
   return Physics::Vector2(x - f, y - f);
 }
 
@@ -31,7 +31,7 @@ Physics::Vector2 Physics::Vector2::operator-(const Physics::Vector2 &v) const {
   return Physics::Vector2(x - v.x, y - v.y);
 }
 
-Physics::Vector2 Physics::Vector2::operator*(const float f) const {
+Physics::Vector2 Physics::Vector2::operator*(const int f) const {
   return Physics::Vector2(x * f, y * f);
 }
 
@@ -39,7 +39,7 @@ Physics::Vector2 Physics::Vector2::operator*(const Physics::Vector2 &v) const {
   return Physics::Vector2(x * v.x, y * v.y);
 }
 
-Physics::Vector2 Physics::Vector2::operator/(const float f) const {
+Physics::Vector2 Physics::Vector2::operator/(const int f) const {
   return Physics::Vector2(x / f, y / f);
 }
 
@@ -48,7 +48,7 @@ void Physics::Vector2::operator+=(const Physics::Vector2 &v) {
   y += v.y;
 }
 
-void Physics::Vector2::operator-=(const float f) {
+void Physics::Vector2::operator-=(const int f) {
   x -= f;
   y -= f;
 }
@@ -58,7 +58,7 @@ void Physics::Vector2::operator-=(const Physics::Vector2 &v) {
   y -= v.y;
 }
 
-void Physics::Vector2::operator*=(const float &f) {
+void Physics::Vector2::operator*=(const int &f) {
   x *= f;
   y *= f;
 }
@@ -72,11 +72,13 @@ Physics::Vector2 Physics::Vector2::clone() const {
   return Physics::Vector2(x, y);
 }
 
-float Physics::Vector2::magnitude() const { return std::sqrt(x * x + y * y); }
+int Physics::Vector2::magnitude() const {
+  return static_cast<int>(std::sqrt(x * x + y * y));
+}
 
 void Physics::Vector2::normalize() {
-  float mag = magnitude();
-  if (mag > 0.0f) {
+  int mag = magnitude();
+  if (mag > 0) {
     x /= mag;
     y /= mag;
   }
@@ -89,14 +91,14 @@ Physics::Vector2 Physics::Vector2::normalized() const {
 }
 
 bool Physics::checkCollision(const Rect &rect1, const Rect &rect2) {
-  float r1x = rect1.p1.x;
-  float r1y = rect1.p1.y;
-  float r1w = rect1.width();
-  float r1h = rect1.height();
-  float r2x = rect2.p1.x;
-  float r2y = rect2.p1.y;
-  float r2w = rect2.width();
-  float r2h = rect2.height();
+  int r1x = rect1.p1.x;
+  int r1y = rect1.p1.y;
+  int r1w = rect1.width();
+  int r1h = rect1.height();
+  int r2x = rect2.p1.x;
+  int r2y = rect2.p1.y;
+  int r2w = rect2.width();
+  int r2h = rect2.height();
 
   // are the sides of one rectangle touching the other?
